anyway_requirements.txt
# Anyway SDK Dependencies for Sandbox Integration
# Install with: pip install -r anyway_requirements.txt

# Core Anyway SDK
traceloop-sdk>=0.51.0

# OpenTelemetry dependencies (usually auto-installed but ensuring compatibility)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0

# Additional dependencies for enhanced tracing
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-asyncio>=0.41b0

# Environment management
python-dotenv>=1.0.0

# HTTP requests for connectivity testing
requests>=2.31.0